refac(HomePage): introducing the HomePage folder to wrap the home page content and refactoring all the file path
