+++ src/App/App.module.css
@@ -1,9 +1,3 @@
 .App {
   background-color: rgba(0,0,0,0.08);
 }
-
-.pageContent {
-  width: 80%;
-  margin: 0 auto;
-  min-height: 75vh;
-}
