--- src/App/components/index.ts
@@ -1,5 +1,0 @@
-export * from './Footer';
-export * from './Header';
-export * from './SideMenu';
-export * from './Carousal';
-export * from './ProductCollection';
