+++ src/pages/HomePage/HomePage.module.scss
@@ -1,0 +1,5 @@
+.pageContent {
+  width: 80%;
+  margin: 0 auto;
+  min-height: 75vh;
+}
