+++ src/pages/HomePage/components/Carousal/Carousal.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import styles from './Carousal.module.css';
 import { Image, Carousel as AntCarousel } from 'antd';
 
-import carouselImage1 from '../../../assets/images/carousel_1.jpg';
-import carouselImage2 from '../../../assets/images/carousel_2.jpg';
-import carouselImage3 from '../../../assets/images/carousel_3.jpg';
+import carouselImage1 from '../../../../assets/images/carousel_1.jpg';
+import carouselImage2 from '../../../../assets/images/carousel_2.jpg';
+import carouselImage3 from '../../../../assets/images/carousel_3.jpg';
 
 export const Carousal: React.FC = () => {
   return(
