+++ src/pages/HomePage/components/Carousal/index.ts
