+++ src/pages/HomePage/components/index.ts
@@ -1,0 +1,3 @@
+export * from './Carousal';
+export * from './ProductCollection';
+export * from './SideMenu';
